+++ app/(tabs)/_layout.tsx
@@ -39,7 +39,7 @@ export default function TabLayout() {
         }}
       />
       <Tabs.Screen
-        name="index"
+        name="notice"
         options={{
           title: "공지",
           tabBarIcon: ({ color }) => (
@@ -48,7 +48,7 @@ export default function TabLayout() {
         }}
       />
       <Tabs.Screen
-        name="calendar"
+        name="index"
         options={{
           title: "과행사",
           tabBarIcon: ({ color }) => (
