+++ app/(tabs)/notice/anotherEvent.tsx
@@ -6,7 +6,7 @@ import { Entypo } from "@expo/vector-icons";
 import { useRouter } from "expo-router";
 import { Pressable, Text } from "react-native";
 
-export default function MajorNoticeScreen() {
+export default function AnotherEventScreen() {
   const router = useRouter();
   return (
     <NormalScrollView
@@ -15,9 +15,8 @@ export default function MajorNoticeScreen() {
           backgroundColor="#fafafa"
           left={
             <Pressable
-              //onPress={() => router.back()}
-              onPress={() => router.push("/(tabs)")}
-              className="flex-row gap-2 items-center"
+              onPress={() => router.back()}
+              className="flex-row items-center"
             >
               <Entypo name="chevron-left" size={30} color="#3a3a3a" />
               <Text className="font-extrabold text-3xl">공지</Text>
@@ -26,11 +25,13 @@ export default function MajorNoticeScreen() {
         />
       }
     >
-      <NoticeTitle title="학부 공지 확인하기"></NoticeTitle>
+      <NoticeTitle title="타 주최 행사 확인하기"></NoticeTitle>
       <NoticeCard
         title="1학기 개강총회"
         Date="2025.03.01"
-        onPress={() => router.push("/(notice)/detail")}
+        place="미쳐버린 파닭"
+        dDay="D-3"
+        onPress={() => router.push("/(tabs)/notice/detail")}
       />
     </NormalScrollView>
   );
