+++ app/(tabs)/notice/majorEvent.tsx
@@ -6,7 +6,7 @@ import { Entypo } from "@expo/vector-icons";
 import { useRouter } from "expo-router";
 import { Pressable, Text } from "react-native";
 
-export default function AnotherEventScreen() {
+export default function MajorEventScreen() {
   const router = useRouter();
   return (
     <NormalScrollView
@@ -15,7 +15,8 @@ export default function AnotherEventScreen() {
           backgroundColor="#fafafa"
           left={
             <Pressable
-              onPress={() => router.back()}
+              //onPress={() => router.back()}
+              onPress={() => router.push("/(tabs)/notice")}
               className="flex-row items-center"
             >
               <Entypo name="chevron-left" size={30} color="#3a3a3a" />
@@ -25,13 +26,13 @@ export default function AnotherEventScreen() {
         />
       }
     >
-      <NoticeTitle title="타 주최 행사 확인하기"></NoticeTitle>
+      <NoticeTitle title="과행사 공지 확인하기"></NoticeTitle>
       <NoticeCard
         title="1학기 개강총회"
         Date="2025.03.01"
         place="미쳐버린 파닭"
-        dDay="D-3"
-        onPress={() => router.push("/(notice)/detail")}
+        dDay="D-2"
+        onPress={() => router.push("/(tabs)/notice/detail")}
       />
     </NormalScrollView>
   );
