+++ app/(tabs)/notice/majorNotice.tsx
@@ -6,7 +6,7 @@ import { Entypo } from "@expo/vector-icons";
 import { useRouter } from "expo-router";
 import { Pressable, Text } from "react-native";
 
-export default function MajorEventScreen() {
+export default function MajorNoticeScreen() {
   const router = useRouter();
   return (
     <NormalScrollView
@@ -16,8 +16,8 @@ export default function MajorEventScreen() {
           left={
             <Pressable
               //onPress={() => router.back()}
-              onPress={() => router.push("/(tabs)")}
-              className="flex-row items-center"
+              onPress={() => router.push("/(tabs)/notice")}
+              className="flex-row gap-2 items-center"
             >
               <Entypo name="chevron-left" size={30} color="#3a3a3a" />
               <Text className="font-extrabold text-3xl">공지</Text>
@@ -26,13 +26,11 @@ export default function MajorEventScreen() {
         />
       }
     >
-      <NoticeTitle title="과행사 공지 확인하기"></NoticeTitle>
+      <NoticeTitle title="학부 공지 확인하기"></NoticeTitle>
       <NoticeCard
         title="1학기 개강총회"
         Date="2025.03.01"
-        place="미쳐버린 파닭"
-        dDay="D-2"
-        onPress={() => router.push("/(notice)/detail")}
+        onPress={() => router.push("/(tabs)/notice/detail")}
       />
     </NormalScrollView>
   );
