+++ app/_layout.tsx
@@ -6,8 +6,8 @@ import {
 import { useFonts } from "expo-font";
 import { SplashScreen, Stack } from "expo-router";
 import { StatusBar } from "expo-status-bar";
-import "react-native-gesture-handler";
 import { GestureHandlerRootView } from "react-native-gesture-handler";
+
 import "react-native-reanimated";
 
 import "@/global.css";
